23 October 2023.

+++ 04_ValueAddedCourse/02_BinSearchandGreedy/Array_Balancing.cpp
@@ -1,24 +1,24 @@
 #include <bits/stdc++.h>
 using namespace :: std;
 
-typedef vector<int> vi;
+typedef vector<long long int> vi;
 
 int main(void) {
-    int test; cin >> test;
-    for(int _ = 0; _ < test; _++){
-        int n; cin >> n;
+    long long int test; cin >> test;
+    for(long long int _ = 0; _ < test; _++){
+        long long int n; cin >> n;
         vi A(n), B(n);
 
-        for(int &a: A) cin >> a;
-        for(int &b: B) cin >> b;
+        for(long long int &a: A) cin >> a;
+        for(long long int &b: B) cin >> b;
 
         unsigned long long int r = 0;
-        for(int i = 1; i < n; i++){
-            int x = abs(A[i] - A[i-1]) + abs(B[i] - B[i-1]);
-            int y = abs(B[i] - A[i-1]) + abs(A[i] - B[i-1]);
+        for(long long int i = 1; i < n; i++){
+            long long int x = abs(A[i] - A[i-1]) + abs(B[i] - B[i-1]);
+            long long int y = abs(B[i] - A[i-1]) + abs(A[i] - B[i-1]);
 
             if(x > y){
-                int temp = A[i];
+                long long int temp = A[i];
                 A[i] = B[i];
                 B[i] = temp;
             }
